requirements.txt
requests~=2.25.1
python-dotenv~=0.20.0
fastapi~=0.78.0
SQLAlchemy~=1.4.39
uvicorn~=0.18.1
asyncpg~=0.25.0
gunicorn~=20.1.0
python-dotenv~=0.20.0